requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
amqp==5.0.9
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
botocore==1.23.26
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==5.0.0
celery==5.2.2
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.6
distlib==0.3.4
distro==1.4.0
dj-database-url==0.5.0
Django==3.2.10
django-axes==5.28.0
django-bootstrap-v5==1.0.8
django-common-helpers==0.9.2
django-compat==1.0.15
django-countries==7.2.1
django-crispy-forms==1.13.0
django-cron==0.5.1
django-crontab==0.7.1
django-crum==0.7.9
django-elasticsearch-dsl==7.2.1
django-formtools==2.3
django-htmx==1.6.0
django-ipware==4.0.2
django-mathfilters==1.0.0
django-phonenumber-field==6.0.0
django-rosetta==0.9.8
django-storages==1.11.1
django-timezone-field==4.2.1
django-twilio==0.13.2
Eel==0.12.4
elasticsearch==7.16.2
elasticsearch-dsl==7.4.0
filelock==3.0.12
forex-python==1.8
frozenlist==1.2.0
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==3.10.1
ipinfo==4.2.1
jmespath==0.10.0
kombu==5.2.2
maxminddb==2.2.0
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
phonenumbers==8.12.40
phonenumberslite==8.12.31
platformdirs==2.2.0
polib==1.1.1
prompt-toolkit==3.0.24
psycopg2==2.9.3
psycopg2-binary==2.9.3
py-moneyed==1.2
pycodestyle==2.8.0
pycountry==20.7.3
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3
redis==3.5.3
regex==2021.8.21
requests==2.26.0
s3transfer==0.5.0
shortcountrynames==0.9.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
ssh-import-id==5.10
supervisor==4.1.0
toml==0.10.2
tomli==1.2.1
twilio==6.63.2
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
whichcraft==0.6.1
whitenoise==5.3.0
yarl==1.7.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
